airflow/requirements.txt
requests==2.31.0
kafka-python==2.0.2
snowflake-connector-python==3.14.0